use double instead of float in WPXProperty* stuff

+++ src/conv/html/HtmlDocumentGenerator.cpp
@@ -142,8 +142,8 @@ void HtmlDocumentGenerator::openParagraph(const WPXPropertyList &propList, const
 		if (propList["fo:text-indent"])
 			*m_pOutputStream << "text-indent:" << propList["fo:text-indent"]->getStr().cstr() << ";";
 
-		if (propList["fo:line-height"] && propList["fo:line-height"]->getFloat() != 1.0f)
-			*m_pOutputStream << "line-height:" << propList["fo:line-height"]->getFloat() << ";";
+		if (propList["fo:line-height"] && propList["fo:line-height"]->getDouble() != 1.0f)
+			*m_pOutputStream << "line-height:" << propList["fo:line-height"]->getDouble() << ";";
 		*m_pOutputStream << "\">";
 	}
 }
